ovn-controller: avoid connecting to local ovs bridge
In the iovisor-ovn architecture the ovs deamon is not running, then it is
not possible to connect to the br-int bridge.
This commit disbales that connection to avoid error during the deployment
of openstack instances.

TODO: it is just a probe of concept, it has to be implemented in a clean
and elegant way